+++ packages/backend/src/schema_def.ts
@@ -98,7 +98,9 @@ export const J_SCHEMA_ORDER_DETAILED = {
     shop_id: { type: "number" },
     shop_name: { type: "string" },
     cour_id: { type: "number" },
+    order_value: { type: "number" },
     order_begin_time: { type: "string" },
+    order_destination: { type: "string" },
     order_state: { type: "number" },
   },
   required: [