+++ packages/frontend/src/main.ts
@@ -2,12 +2,12 @@ import { createApp } from "vue";
 import App from "./App.vue";
 
 import "./main.css";
-import '@mdi/font/css/materialdesignicons.css';
+import "@mdi/font/css/materialdesignicons.css";
 import "vuetify/styles";
 import { createVuetify } from "vuetify";
 import * as components from "vuetify/components";
 import * as directives from "vuetify/directives";
-import axios from "axios";
+import axios, { AxiosError, type AxiosResponse } from "axios";
 
 import dayjs from "dayjs";
 import "dayjs/locale/zh-cn";
@@ -26,6 +26,17 @@ axios.interceptors.request.use((config) => {
   }
   return config;
 });
+axios.interceptors.response.use(
+  (response) => response,
+  (err) => {
+    if (err instanceof AxiosError && err.response) {
+      const { data } = err.response as AxiosResponse<{ message: string }>;
+      alert(data.message);
+    }
+    throw err;
+  }
+);
+
 dayjs.extend(relativeTime);
 dayjs.locale("zh-cn");